Add files via upload
The models strongest point is its microphysics -> macrophysics bridge for Dwarf Halos. This PDF explicitly focuses on that aspect and showcases how the HLS one loop accounts for the runaway R∗. Unless a significant coupling exist that causes a tilt in the 2 loop variant the current implementation should create a fit across scales from cusp -> core.

Signed-off-by: Voxtrium <voxtrium@gmail.com>